apicius/pyproject.toml
[tool.poetry]
name = "apicius"
version = "0.1.0"
description = ""
authors = ["Paul Briton"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
django-tailwind = "^3.5.0"
django-htmx = "^1.15.0"
django-extensions = "^3.2.3"
django-environ = "^0.10.0"
humanfriendly = "^10.0"
django-crispy-forms = "^2.0"
crispy-tailwind = "^0.5.0"
python-dotenv = "^1.0.0"
django-modeltranslation = "^0.18.10"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
djhtml = "^3.0.6"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
